+++ R/shinyapp.R
@@ -345,6 +345,17 @@ loadSupport <- function(appDir=NULL, renv=new.env(parent=globalenv()), globalren
     appDir <- findEnclosingApp(".")
   }
 
+  descFile <- file.path.ci(appDir, "DESCRIPTION")
+  if (file.exists(file.path.ci(appDir, "NAMESPACE")) ||
+      (file.exists(descFile) &&
+       identical(as.character(read.dcf(descFile, fields = "Type")), "Package")))
+  {
+    warning(
+      "Loading R/ subdirectory for Shiny application, but this directory appears ",
+      "to contain an R package. Sourcing files in R/ may cause unexpected behavior."
+    )
+  }
+
   if (!is.null(globalrenv)){
     # Evaluate global.R, if it exists.
     globalPath <- file.path.ci(appDir, "global.R")